+++ misc/allow-block-list/src/lib.rs
@@ -354,9 +354,16 @@ mod tests {
             .block_peer(*listener.local_peer_id());
 
         let (
-            [SwarmEvent::ConnectionClosed { peer_id: closed_dialer_peer, .. }],
-            [SwarmEvent::ConnectionClosed { peer_id: closed_listener_peer, .. }]
-        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await else {
+            [SwarmEvent::ConnectionClosed {
+                peer_id: closed_dialer_peer,
+                ..
+            }],
+            [SwarmEvent::ConnectionClosed {
+                peer_id: closed_listener_peer,
+                ..
+            }],
+        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await
+        else {
             panic!("unexpected events")
         };
         assert_eq!(closed_dialer_peer, *listener.local_peer_id());
@@ -417,9 +424,22 @@ mod tests {
             .unwrap();
 
         let (
-            [SwarmEvent::OutgoingConnectionError { error: DialError::Denied { cause: outgoing_cause }, .. }],
-            [_, SwarmEvent::IncomingConnectionError { error: ListenError::Denied { cause: incoming_cause }, .. }],
-        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await else {
+            [SwarmEvent::OutgoingConnectionError {
+                error:
+                    DialError::Denied {
+                        cause: outgoing_cause,
+                    },
+                ..
+            }],
+            [_, SwarmEvent::IncomingConnectionError {
+                error:
+                    ListenError::Denied {
+                        cause: incoming_cause,
+                    },
+                ..
+            }],
+        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await
+        else {
             panic!("unexpected events")
         };
         assert!(outgoing_cause.downcast::<NotAllowed>().is_ok());
@@ -447,9 +467,16 @@ mod tests {
             .list
             .disallow_peer(*listener.local_peer_id());
         let (
-            [SwarmEvent::ConnectionClosed { peer_id: closed_dialer_peer, .. }],
-            [SwarmEvent::ConnectionClosed { peer_id: closed_listener_peer, .. }]
-        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await else {
+            [SwarmEvent::ConnectionClosed {
+                peer_id: closed_dialer_peer,
+                ..
+            }],
+            [SwarmEvent::ConnectionClosed {
+                peer_id: closed_listener_peer,
+                ..
+            }],
+        ) = libp2p_swarm_test::drive(&mut dialer, &mut listener).await
+        else {
             panic!("unexpected events")
         };
         assert_eq!(closed_dialer_peer, *listener.local_peer_id());