+++ transports/quic/CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.9.2
+
+- Cut stable release.
+
+## 0.9.2-alpha
+
+- Add support for reusing an existing socket when dialing localhost address.
+  See [PR 4304].
+
+[PR 4304]: https://github.com/libp2p/rust-libp2p/pull/4304
+
+## 0.9.1-alpha
+
+- Allow listening on ipv4 and ipv6 separately.
+  See [PR 4289].
+
+[PR 4289]: https://github.com/libp2p/rust-libp2p/pull/4289
+
+## 0.9.0-alpha
+
+- Use `quinn` instead of `quinn-proto`.
+  See [PR 3454].
+
+[PR 3454]: https://github.com/libp2p/rust-libp2p/pull/3454
+
 ## 0.8.0-alpha
 
 - Raise MSRV to 1.65.